+++ CotEditor/Sources/Settings Window/Panes/AppearanceSettingsView.swift
@@ -188,7 +188,7 @@ struct AppearanceSettingsView: View {
             }
         }
         .scenePadding()
-        .frame(width: 600)
+        .frame(width: 610)
     }
 }