+++ CotEditor/Sources/Settings Window/Panes/DonationSettingsView.swift
@@ -204,7 +204,7 @@ struct DonationSettingsView: View {
             }
         }
         .scenePadding()
-        .frame(width: 600)
+        .frame(width: 610)
     }